requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.1
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
pyjwt>=2.8.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0.post1
python-multipart>=0.0.9
slowapi>=0.1.9
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.26.0
